+++ src/lib/DataObject/Factory.php
@@ -137,7 +137,7 @@ class Factory
     /**
      * Get all.
      */
-    public function getAll(CollectionInterface $collection, ?array $query = null, bool $include_dataset = true, ?int $offset = 0, ?int $limit = 0, ?array $sort = ['$natural' => -1]): Generator
+    public function getAll(CollectionInterface $collection, ?array $query = null, bool $include_dataset = true, ?int $offset = 0, ?int $limit = 0, ?array $sort = null): Generator
     {
         $that = $this;
 
@@ -222,7 +222,7 @@ class Factory
         $that = $this;
 
         return $this->resource_factory->watchFrom($this->db->{$collection->getCollection()}, $after, $existing, $query, function (array $resource) use ($collection, $that) {
-            return $that->build($collection, $resource);
+            return $that->build($resource, $collection);
         }, $offset, $limit, $sort);
     }