+++ src/lib/Helper.php
@@ -16,7 +16,7 @@ class Helper
     /**
      * Propper json_decode.
      */
-    public static function jsonDecode($data, $options)
+    public static function jsonDecode(string $data, bool $options = true)
     {
         $result = json_decode($data, $options);
 
@@ -85,10 +85,6 @@ class Helper
      */
     public static function associativeArrayToPath(iterable $arr, iterable $narr = [], $nkey = ''): array
     {
-        /*if ($nkey !== '') {
-            $narr[substr($nkey, 0, -1)] = $arr;
-        }*/
-
         foreach ($arr as $key => $value) {
             if (is_array($value)) {
                 $narr = array_merge($narr, self::associativeArrayToPath($value, $narr, $nkey.$key.'.'));